requirements.txt
# AI Bootcamp Tutor MVP - Backend Dependencies
# Python 3.10+ required

# Flask Framework
Flask==2.3.3
flask-cors==4.0.0

# Environment Management
python-dotenv==1.0.0

# Database & ORM
prisma==0.11.0
psycopg2-binary==2.9.7

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Async Support
asyncio-cors==0.1.1

# Production Server
gunicorn==21.2.0

# Logging & Monitoring
structlog==23.1.0

# Development & Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1
black==23.7.0
flake8==6.0.0

# Security
cryptography==41.0.4

# Utilities
requests==2.31.0
python-dateutil==2.8.2
